+++ apps-client/mobile/app/(tabs)/(feed,official,profile)/election/[electionId]/index.tsx
@@ -16,6 +16,7 @@ import { AvatarPPLEFallback } from '@app/components/avatar-pple-fallback'
 import { ElectionDetailCard } from '@app/components/election/election-card'
 import { SafeAreaLayout } from '@app/components/safe-area-layout'
 import { reactQueryClient } from '@app/libs/api-client'
+import { useAuthMe } from '@app/libs/auth'
 import { createImageUrl } from '@app/utils/image'
 
 export default function ElectionDetailPage() {
@@ -81,14 +82,15 @@ export default function ElectionDetailPage() {
 }
 
 function ElectionAction({ election }: { election: ElectionWithCurrentStatus }) {
+  const user = useAuthMe()
   if (
     election.status === 'OPEN_VOTE' &&
     (election.type === 'ONLINE' || election.type === 'HYBRID')
   ) {
     return (
       <View className="p-4">
-        <Link href={`/election/${election.id}/vote`} asChild>
-          <Button>
+        <Link disabled={user.data?.isSuspended} href={`/election/${election.id}/vote`} asChild>
+          <Button disabled={user.data?.isSuspended}>
             <Icon icon={VoteIcon} size={16} />
             <Text>{election.isVoted ? 'ลงคะแนนใหม่' : 'ลงคะแนน'}</Text>
           </Button>
@@ -124,17 +126,19 @@ function ElectionCandidateList({ election }: { election: GetElectionResponse }) 
       <H2 className="text-xs font-heading-semibold text-base-text-high mb-1">
         รายชื่อผู้ลงสมัครเลือกตั้ง
       </H2>
-      <View className="flex flex-col gap-3">
+      <View className="flex flex-col">
         {election.candidates.map((candidate) => (
-          <View key={candidate.id} className="py-3 px-1 flex flex-row gap-2 items-center">
-            <View className="flex flex-col gap-1 items-center">
-              <Text className="text-xs font-heading-semibold text-base-primary-default -mb-1">
-                เบอร์
-              </Text>
-              <Text className="text-3xl font-heading-bold text-base-primary-default">
-                {candidate.number}
-              </Text>
-            </View>
+          <View key={candidate.id} className="py-1 px-1 flex flex-row gap-2 items-center">
+            {candidate.number !== null && (
+              <View className="flex flex-col gap-1 items-center">
+                <Text className="text-xs font-heading-semibold text-base-primary-default -mb-1">
+                  เบอร์
+                </Text>
+                <Text className="text-3xl font-heading-bold text-base-primary-default">
+                  {candidate.number}
+                </Text>
+              </View>
+            )}
             <Avatar alt="Candidate Profile Image" className="size-10">
               {election.candidates[0].profileImageUrl && (
                 <AvatarImage