+++ packages/database/prisma/migrations/20251208155219_add_order_to_mini_app/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "public"."MiniApp" ADD COLUMN     "order" INTEGER NOT NULL DEFAULT 0;